refactor: move the generic files
Move the files into the top level series package since they can be shared
