+++ app/src/main/java/com/chesire/malime/flow/series/list/SeriesViewHolder.kt
@@ -1,4 +1,4 @@
-package com.chesire.malime.flow.series.list.anime
+package com.chesire.malime.flow.series.list
 
 import android.view.View
 import androidx.core.view.ViewCompat
